+++ src/main/java/net/pwall/util/MiniSet5.java
@@ -27,6 +27,7 @@ package net.pwall.util;
 
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.Set;
 
 /**
@@ -61,6 +62,33 @@ public class MiniSet5<T> extends MiniSet<T> {
     }
 
     /**
+     * Construct a {@code MiniSet5} from another {@link Set} (helps with deserializing).
+     *
+     * @param   set         the other {@link Set}
+     * @throws  IllegalArgumentException if the size of the other set is not 5
+     */
+    public MiniSet5(Set<T> set) {
+        if (set.size() != 5)
+            throw new IllegalArgumentException("MiniSet5 size must be 5");
+        if (set instanceof MiniSet5) {
+            MiniSet5<T> miniSet5 = (MiniSet5<T>)set;
+            value0 = miniSet5.value0;
+            value1 = miniSet5.value1;
+            value2 = miniSet5.value2;
+            value3 = miniSet5.value3;
+            value4 = miniSet5.value4;
+        }
+        else {
+            Iterator<T> iterator = set.iterator();
+            value0 = iterator.next();
+            value1 = iterator.next();
+            value2 = iterator.next();
+            value3 = iterator.next();
+            value4 = iterator.next();
+        }
+    }
+
+    /**
      * Get the number of values (always five).
      *
      * @return      the number of values
@@ -88,7 +116,8 @@ public class MiniSet5<T> extends MiniSet<T> {
      */
     @Override
     public boolean contains(Object o) {
-        return value0.equals(o) || value1.equals(o) || value2.equals(o) || value3.equals(o) || value4.equals(o);
+        return o == null ? value0 == null || value1 == null || value2 == null || value3 == null || value4 == null :
+            o.equals(value0) || o.equals(value1) || o.equals(value2) || o.equals(value3) || o.equals(value4);
     }
 
     /**
@@ -161,7 +190,8 @@ public class MiniSet5<T> extends MiniSet<T> {
      */
     @Override
     public int hashCode() {
-        return value0.hashCode() + value1.hashCode() + value2.hashCode() + value3.hashCode() + value4.hashCode();
+        return Objects.hashCode(value0) + Objects.hashCode(value1) + Objects.hashCode(value2) +
+                Objects.hashCode(value3) + Objects.hashCode(value4);
     }
 
     /**